+++ clinic/src/app/core/auth/auth.interceptor.ts
@@ -1,0 +1,18 @@
+// auth.interceptor.ts
+import { Injectable } from '@angular/core';
+import { HttpInterceptor, HttpHandler, HttpRequest } from '@angular/common/http';
+import { AuthserviceService } from './authservice.service';
+
+@Injectable()
+export class AuthInterceptor implements HttpInterceptor {
+  constructor(private authService: AuthserviceService) {}
+
+  intercept(req: HttpRequest<any>, next: HttpHandler) {
+    const authToken = this.authService.getToken().token;
+    // console.log('AuthToken:', authToken);
+    const authReq = req.clone({
+      setHeaders: { Authorization: `Bearer ${authToken}` }
+    });
+    return next.handle(authReq);
+  }
+}
