+++ clinic/src/app/registration/registration.component.ts
@@ -1,5 +1,5 @@
 import { Component, ViewChild } from '@angular/core';
-import { AuthserviceService } from '../shared/authservice.service';
+import { AuthserviceService } from '../core/auth/authservice.service';
 import { HttpErrorResponse } from '@angular/common/http';
 import { Router } from '@angular/router';
