+++ clinic/src/environments/environment.development.ts
@@ -1,0 +1,4 @@
+export const environment = {
+    production: false,
+    apiUrl: 'http://localhost:5100/api',
+  };
